Changes:
- prepare tests for environment updates (drop_database)
- improve/clarify settings of some tests
- document in TODOs where we should check for hidden load calls

+++ tests/models/test_model_class.py
@@ -37,7 +37,8 @@ async def create_test_database():
     async with database:
         await models.create_all()
         yield
-    await models.drop_all()
+        if not database.drop:
+            await models.drop_all()
 
 
 def test_model_class():
