+++ tests/models/test_model_multiple_primary_keys.py
@@ -26,7 +26,8 @@ async def create_test_database():
     async with database:
         await models.create_all()
         yield
-        await models.drop_all()
+        if not database.drop:
+            await models.drop_all()
 
 
 async def test_model_multiple_primary_key():
